--- a/app/models/assessment.rb
+++ b/app/models/assessment.rb
@@ -8,7 +8,17 @@ class Assessment < ApplicationRecord
   belongs_to :country, foreign_key: "country_alpha3", primary_key: "alpha3"
   has_many :scores, class_name: "AssessmentScore"
   has_many :plans
-  default_scope { includes(:country).order("country_alpha3") }
+  default_scope { order("country_alpha3") }
+
+  scope :with_publication, -> (country_alpha3, assessment_publication_id) {
+    includes(:assessment_publication)
+      .find_by_country_alpha3_and_assessment_publication_id(country_alpha3, assessment_publication_id)
+  }
+
+  scope :deep_load, -> (country_alpha3, assessment_publication_id) {
+    includes({scores: :assessment_indicator})
+      .with_publication(country_alpha3, assessment_publication_id)
+  }
 
   delegate :jee1?, :spar_2018?, :type_description, to: :assessment_publication