--- a/app/models/plan_goal.rb
+++ b/app/models/plan_goal.rb
@@ -2,6 +2,4 @@ class PlanGoal < ApplicationRecord
   belongs_to :plan
   belongs_to :assessment_indicator
   belongs_to :benchmark_indicator
-
-  default_scope { includes(:assessment_indicator) }
 end