--- a/app/views/pages/home.html.haml
+++ b/app/views/pages/home.html.haml
@@ -22,5 +22,5 @@
             analysis of your actions.
       .row.my-4
         .col-8
-          = link_to "Get Started", get_started_url, class: "btn btn-success", type: "button"
+          = link_to "Get Started", get_started_url, class: "btn btn-success cta-btn", style: "width: 250px", type: "button"
     .col.right-panel.m-auto= render "hero_image"