--- a/app/views/plans/assessment_not_found.html.haml
+++ b/app/views/plans/assessment_not_found.html.haml
@@ -1,3 +1,3 @@
 .col
   %h2
-    There is no "#{@assessment_type}" assessment available for #{@country_name}.
+    There is no "#{params[:assessment_type]}" assessment available for #{params[:country_name]}.