--- a/app/views/plans/index.html.erb
+++ b/app/views/plans/index.html.erb
@@ -4,7 +4,7 @@
       <h1>WELCOME</h1>
     </div>
     <div class="col-2 pl-0 py-4">
-      <%= button_to "CREATE PLAN", get_started_url, class: "btn btn-primary btn-block", type: "button", method: :get %>
+      <%= link_to "CREATE PLAN", get_started_url, class: "btn btn-success cta-btn", type: "button", style: "width: 220px" %>
     </div>
   </div>