--- a/db/migrate/20200202214510_add_more_ids_to_plan_activity_for_performance.rb
+++ b/db/migrate/20200202214510_add_more_ids_to_plan_activity_for_performance.rb
@@ -1,0 +1,8 @@
+class AddMoreIdsToPlanActivityForPerformance < ActiveRecord::Migration[5.2]
+  def change
+    add_column :plan_activities, :benchmark_indicator_id, :integer
+    add_column :plan_activities, :benchmark_technical_area_id, :integer
+    add_foreign_key :plan_activities, :benchmark_indicators
+    add_foreign_key :plan_activities, :benchmark_technical_areas
+  end
+end