--- a/test/models/plan_test.rb
+++ b/test/models/plan_test.rb
@@ -334,13 +334,48 @@ describe Plan do
   end
 
   describe "#count_activities_by_ta" do
-    let(:plan) { create(:plan_nigeria_jee1) }
+    describe "for a full plan" do
+      let(:benchmark_technical_areas) { BenchmarkTechnicalArea.all }
+      let(:plan) { create(:plan_nigeria_jee1) }
 
-    it "returns an array of the expected integers" do
-      expected = [6, 12, 19, 9, 11, 13, 19, 7, 15, 18, 11, 15, 7, 19, 20, 16, 14, 4]
-      plan.count_activities_by_ta.must_equal expected
+      it "returns an array of the expected integers" do
+        expected = [6, 12, 19, 9, 11, 13, 19, 7, 15, 18, 11, 15, 7, 19, 20, 16, 14, 4]
+        plan.count_activities_by_ta(benchmark_technical_areas).must_equal expected
+      end
+    end
+
+    describe "for a plan of sparsely populated technical areas" do
+      let(:benchmark_technical_areas) { BenchmarkTechnicalArea.all }
+      let(:indicator_attrs) do
+        {
+            jee1_ind_p21: "2",
+            jee1_ind_p21_goal: "3",
+            jee1_ind_d21: "3",
+            jee1_ind_d21_goal: "4",
+            jee1_ind_d22: "2",
+            jee1_ind_d22_goal: "3",
+            jee1_ind_d23: "3",
+            jee1_ind_d23_goal: "4",
+            jee1_ind_d24: "3",
+            jee1_ind_d24_goal: "4",
+        }.with_indifferent_access
+      end
+      let(:plan) do
+        Plan.create_from_goal_form(
+          indicator_attrs: indicator_attrs,
+          assessment: assessment_for_nigeria_jee1,
+          is_5_year_plan: true,
+          plan_name: "test plan 3737"
+        )
+      end
+
+      it "returns an array of the expected integers" do
+        expected = [0, 5, 0, 0, 0, 0, 0, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0]
+        plan.count_activities_by_ta(benchmark_technical_areas).must_equal expected
+      end
     end
   end
+
 
   describe "#activities_for" do
     let(:plan) { create(:plan_nigeria_jee1) }