… Gemfile.lock
… test/fixtures/files/plan_for_nigeria_jee1/plan_activities.json