- Plan#create for JEE 1.0 1-year plan went from about 4 sec to about 2 sec. Plan#show for same went from about 6 sec to about 1 sec.
- gains PlanBuilder#create_from_goal_form are due to fixing what could be called a bug where the Plan.transaction was called more than once. Also from accessing IDs instead of loading the object
- gains on Plan#show are from putting more IDs on the plan_activities table and using those in the loops that connect together data
- modify plan#count_activities_by_ta to work in a more performant way, taking advantage of the schema change to tally the activity count with fewer hops across object associations
- on advise from the bullet gem made changes to better optimize queries for their uses by templates/etc, made use of custom scopes/finders
- add new class BenchmarkDocument to encapsulate Benchmarks and all its related data: technical areas, indicators, activities
- update the plan_activities.json file and its README per the schema changes
- there is a system test assertion failing for no apparent reason